refactor: rename package acme to lego.
