+++ lego/client_config.go
@@ -1,4 +1,4 @@
-package acme
+package lego
 
 import (
 	"crypto/tls"
