+++ lego/client_test.go
@@ -1,4 +1,4 @@
-package acme
+package lego
 
 import (
 	"crypto"
